LPD-93368 Extract common test logic to BaseObjectActionExecutorTestCase
